Namespace the form's components under Admin::OrganizationForm
The three components become Admin::OrganizationForm::Wrapper,
::FeatureSettings and ::LocationFields, so admin/organization_form is a
namespace folder rather than a component with siblings alongside it -
matching org/search and registrations/show.

Also drops the location fields' TODO about using
organized/manages/location_fields instead. That partial has two render
modes, is localized, and is wired to the manage_locations CoffeeScript
through eight class hooks the admin page doesn't load; the part worth
sharing is the address block, which both already render through
Org::LocationAddressFields.
