+++ app/views/admin/organizations/edit.html.haml
@@ -1,7 +1,7 @@
 = render partial: "/admin/organizations/display_header", locals: {}
 
 = form_for [:admin, @organization], html: {data: {controller: "admin--organization-form"}} do |f|
-  = render(Admin::OrganizationForm::Component.new(form_builder: f, organization: @organization, current_user:, embedable_email: @embedable_email))
+  = render(Admin::OrganizationForm::Wrapper::Component.new(form_builder: f, organization: @organization, current_user:, embedable_email: @embedable_email))
 
   .mt-4
     = f.submit "Update", class: 'btn-lg btn-success float-right'
