+++ app/views/admin/organizations/new.html.haml
@@ -2,7 +2,7 @@
 
 
 = form_for [:admin, @organization], html: {data: {controller: "admin--organization-form"}} do |f|
-  = render(Admin::OrganizationForm::Component.new(form_builder: f, organization: @organization, current_user:))
+  = render(Admin::OrganizationForm::Wrapper::Component.new(form_builder: f, organization: @organization, current_user:))
 
   .row
     .col-sm-10
