+++ spec/components/admin/organization_form/wrapper/component_spec.rb
@@ -2,14 +2,14 @@
 
 require "rails_helper"
 
-RSpec.describe Admin::OrganizationForm::Component, type: :component do
+RSpec.describe Admin::OrganizationForm::Wrapper::Component, type: :component do
   let(:organization) { FactoryBot.create(:organization) }
   let(:current_user) { FactoryBot.create(:superuser) }
 
   def rendered_component(organization, current_user)
     render_in_view_context do
       form_for [:admin, organization] do |f|
-        render(Admin::OrganizationForm::Component.new(form_builder: f, organization:, current_user:))
+        render(Admin::OrganizationForm::Wrapper::Component.new(form_builder: f, organization:, current_user:))
       end
     end
   end
